display total score in endgame menu
