+++ index.html
@@ -20,6 +20,7 @@
     <h2>Game Over</h2>
     <div id="restart-button"><p>PLAY AGAIN</p></div>
     <br>
+
     <div id="score-form">
       Game score: <span id="game-score">0</span>
       <br><br>
@@ -32,6 +33,7 @@
       </span>
     </div>
     <br>
+
     <div id="top-scores">
       <!--> see list of scores; when works show top five scores -->
     </div>
