+++ lib/draw.js
@@ -100,7 +100,7 @@ Draw.prototype.drawScore = function(){
 
 Draw.prototype.checkEnd = function() {
   if(this.qbert.lives === 0) {
-    var endGame = new EndGame();
+    var endGame = new EndGame(this.score);
     endGame.end();
   } else {
     var self = this;
