+++ lib/end_game.js
@@ -1,11 +1,16 @@
 var $ = require('jQuery');
+var Standings = require('./standings');
 
-function EndGame() {
+function EndGame(score) {
   this.endMenu = $("#end-menu");
+  this.score = score;
+  this.standings = new Standings();
 }
 
 EndGame.prototype.end = function(){
+  $('#game-score').text(this.score.total);
   this.endMenu.fadeIn('slow', function(){clearCharacter();});
+  this.standings.display();
 };
 
 function clearCharacter(){
@@ -14,5 +19,15 @@ function clearCharacter(){
   context.clearRect(0, 0, canvas.width, canvas.height);
 }
 
+document.getElementById('initials-button').addEventListener('click', function(){
+  updateStandings();
+}, false);
+
+
+EndGame.prototype.updateStandings = function(){
+  var player = $('#initials').text();
+  var score = $('#game-score').text();
+  this.standings.addGame(player, totalScore);
+}
 
 module.exports = EndGame;
