+++ lib/game.js
@@ -4,7 +4,6 @@ var Ball        = require('./ball');
 var UserInput   = require('./user_input');
 var Draw        = require('./draw');
 var Score       = require('./score');
-var Scoreboard  = require('./scoreboard');
 
 function Game() {
   this.bgCanvas    = document.getElementById('bgCanvas');
@@ -36,7 +35,6 @@ Game.prototype.setInput = function() {
 Game.prototype.startGame = function() {
   this.board.initializeCubes();
   this.board.drawCubes();
-  this.getScoreboard();
   var thisDraw = this.draw;
   window.requestAnimationFrame(function(){thisDraw.drawFrame();});
 };
@@ -107,10 +105,6 @@ Game.prototype.resetDifficulty = function() {
 Game.prototype.resetTick = function(){
   this.draw.tick = 0;
 };
-
-Game.prototype.getScoreboard = function(){
-  this.scoreboard = new Scoreboard();
-}
 
 Game.prototype.resetScoreboard = function(){
   var scoreDiv = document.getElementById("scoreboard");
