+++ lib/index.js
@@ -6,6 +6,7 @@ var startMenu     = $("#start-menu");
 var endMenu       = $("#end-menu");
 var startButton   = document.getElementById("start-button");
 var restartButton = document.getElementById("restart-button");
+var initialsButton = document.getElementById("initials-button");
 
 startButton.addEventListener('click', function(){
   startMenu.fadeOut('slow');
@@ -17,3 +18,7 @@ restartButton.addEventListener('click', function() {
   game.resetGame();
   endMenu.fadeOut('slow');
 });
+
+initialsButtons.addEventListener('click', function(){
+
+})
